move ArtifactAppReference class to internal.io.reference package
