+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/ArtifactComponentReference.java
@@ -25,6 +25,7 @@ import org.wso2.carbon.uuf.api.reference.FragmentReference;
 import org.wso2.carbon.uuf.api.reference.LayoutReference;
 import org.wso2.carbon.uuf.api.reference.PageReference;
 import org.wso2.carbon.uuf.exception.FileOperationException;
+import org.wso2.carbon.uuf.internal.io.reference.ArtifactAppReference;
 
 import java.io.IOException;
 import java.nio.file.Files;
