+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/ArtifactThemeReference.java
@@ -21,6 +21,7 @@ package org.wso2.carbon.uuf.internal.io;
 import org.wso2.carbon.uuf.api.reference.FileReference;
 import org.wso2.carbon.uuf.api.reference.ThemeReference;
 import org.wso2.carbon.uuf.exception.FileOperationException;
+import org.wso2.carbon.uuf.internal.io.reference.ArtifactAppReference;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
